% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Simulation_AdaptiveRandomisationmethodRatioCalc.R
\name{ARmethod}
\alias{ARmethod}
\title{This function adjusts the posterior randomisation probability for each arm using many approachs.
Currently Thall's approach and Trippa's approach are used.
Double biased coin and other method will be added in the next version.}
\usage{
ARmethod(
  Fixratio,
  BARmethod,
  group,
  stats,
  post.prob.btcontrol,
  K,
  n,
  tuningparameter,
  c,
  post.prob.best,
  max.ar,
  armleft,
  treatmentindex
)
}
\arguments{
\item{Fixratio}{A indicator TRUE/FALSE}

\item{BARmethod}{The indicator of which adaptive randomisation method is used}

\item{group}{The current stage}

\item{stats}{The output matrix}

\item{post.prob.btcontrol}{The vector of posterior probability of each active treatment arm better than control}

\item{K}{Total number of arms at the beginning}

\item{n}{The vector of sample size for each arm}

\item{tuningparameter}{The tuning parameter indicator for Thall's approach}

\item{c}{The tuning parameter for Thall's approach}

\item{post.prob.best}{Posterior probability of each arm to be the best}

\item{max.ar}{The upper boundary for randomisation ratio for each arm}

\item{armleft}{The number of treatment left in the platform (>2)}

\item{treatmentindex}{The vector of treatment arm index excluding the control arm whose index is 0}
}
\value{
randomprob: The vector of adjusted randomisation probability to each arm
}
\description{
This function adjusts the posterior randomisation probability for each arm using many approachs.
Currently Thall's approach and Trippa's approach are used.
Double biased coin and other method will be added in the next version.
}
