% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Boundaryconstruction.R
\name{Boundaryconstruction}
\alias{Boundaryconstruction}
\title{This function constructs the stopping boundary based on input information}
\usage{
Boundaryconstruction(Stopbound.inf = Stopbound.inf, ns = ns)
}
\arguments{
\item{Stopbound.inf}{The list of stop boundary information}

\item{ns}{Vector of number of patient at each stage}
}
\value{
The futility boundary and the efficacy boundary
}
\description{
This function constructs the stopping boundary based on input information
}
\examples{
 Stopbound.inf=list(Stop.type="Early-Pocock",Boundary.type="Symmetric",cutoff=c(0.9928,0.0072))
 ns=c(60,120,180,240,300)

 Boundaryconstruction(Stopbound.inf, ns)
 ------------------------------------------
}
