% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Parameterinitialisation.R
\name{Initializetrialparameter}
\alias{Initializetrialparameter}
\title{This function initialises the inner parameter used in simulate.trial function}
\usage{
Initializetrialparameter(response.probs, ns)
}
\arguments{
\item{response.probs}{A vector of response probability of each arm}

\item{ns}{A vector of number of patient at each stage}
}
\value{
A list of initialised parameters
}
\description{
This function initialises the inner parameter used in simulate.trial function
}
\examples{
Initializetrialparameter(response.probs = c(0.4,0.6), ns = c(15,30,45,60,75,90))

#$K
#[1] 2

#$armleft
#[1] 2

#$treatmentindex
#[1] 1

#$n
#[1] 0 0

#$y1
#[1] 0 0

#$groupwise.response.probs
#    [,1] [,2]
#[1,]  0.4  0.6
#[2,]  0.4  0.6
#[3,]  0.4  0.6
#[4,]  0.4  0.6
#[5,]  0.4  0.6
#[6,]  0.4  0.6

#$randomprob
#    1   2
#[1,] 0.5 0.5

#$z
#numeric(0)

#$y
#numeric(0)

#$group_indicator
#numeric(0)

}
