% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Mixeffectmodelanalysis.R
\name{Mixeffect_analysis}
\alias{Mixeffect_analysis}
\title{Jags mixeffect (Time machine) model analysis}
\usage{
Mixeffect_analysis(postsamp.list, group, treatmentindex, ns, K = K)
}
\arguments{
\item{postsamp.list}{A jag result list}

\item{group}{A numerical value indicating the current stage index}

\item{treatmentindex}{The vector of current active arms' index}

\item{ns}{A vector of accumulated number of patient at each stage}

\item{K}{Total number of arms}
}
\value{
A list of evaluation metrics
}
\description{
This function analysis the posterior samples from time machine model
}
\examples{
jagmodel<-"  model {
for(i in 1:armleft){
  for(j in 1:stage){
      Y[i,j] ~ dbin(p[i,j],N[i,j])
          logit(p[i,j]) = beta0 + alpha[stage-(j-1)] + beta1[i]
            }
            }
       alpha[1] = 0
       alpha[2] ~ dnorm(0, tau2)
   for(k in 3:stage ) {
     alpha[k] ~ dnorm(2*alpha[k-1] - alpha[k-2],tau2)
     }
   beta1[1] <- 0
   for(i in 2:armleft){
     beta1[i] ~ dnorm(0,0.31)
     }
   tau2 ~ dgamma(0.1, 0.01)
   beta0 ~ dnorm(0, 0.31)
   }"
   ntemp=matrix(rep(NA,5*2),nrow=2)
   ytemp=matrix(rep(NA,5*2),nrow=2)
   ytemp[1,1:2]=c(5,6);ytemp[2,1:2]=c(5,5)
   ntemp[1,1:2]=c(30,30);ntemp[2,1:2]=c(30,30)
postsamp.list=Mixeffect_modelling(ytemp=ytemp,
treatmentindex=1, group=2, ntemp=ntemp,
armleft=2, jagmodel=jagmodel)
analysis=Mixeffect_analysis(postsamp.list=postsamp.list,
group=2, treatmentindex=2, ns = c(60,120,180,240,300), K=2)
}
