% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Mixeffectmodelling.R
\name{Mixeffect_modelling}
\alias{Mixeffect_modelling}
\title{Randomeffect modelling function}
\usage{
Mixeffect_modelling(ytemp, treatmentindex, group, ntemp, armleft, jagmodel)
}
\arguments{
\item{ytemp}{The data matrix including the information of total number of patients outcomes for each arm at each stage}

\item{treatmentindex}{The vector of current active arms' index}

\item{group}{A numerical value indicating the current stage index}

\item{ntemp}{The data matrix including the information of total number of patients allocated to each arm at each stage}

\item{armleft}{A numerical value indicating the number of active arms and control}

\item{jagmodel}{The jag time machine model}
}
\value{
A dataframe of posterior samples for each parameters
}
\description{
This functions use rjags to sample from the time machine model
}
\examples{
jagmodel<-"  model {
for(i in 1:armleft){
  for(j in 1:stage){
      Y[i,j] ~ dbin(p[i,j],N[i,j])
          logit(p[i,j]) = beta0 + alpha[stage-(j-1)] + beta1[i]
            }
            }
       alpha[1] = 0
       alpha[2] ~ dnorm(0, tau2)
   for(k in 3:stage ) {
     alpha[k] ~ dnorm(2*alpha[k-1] - alpha[k-2],tau2)
     }
   beta1[1] <- 0
   for(i in 2:armleft){
     beta1[i] ~ dnorm(0,0.31)
     }
   tau2 ~ dgamma(0.1, 0.01)
   beta0 ~ dnorm(0, 0.31)
   }"
   ntemp=matrix(rep(NA,5*2),nrow=2)
   ytemp=matrix(rep(NA,5*2),nrow=2)
   ytemp[1,1:2]=c(5,6);ytemp[2,1:2]=c(5,5)
   ntemp[1,1:2]=c(30,30);ntemp[2,1:2]=c(30,30)
postsamp.list=Mixeffect_modelling(ytemp=ytemp,
treatmentindex=1, group=2,
ntemp=ntemp, armleft=2,
jagmodel=jagmodel)
}
