% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Randomisationinformationcheck.R
\name{Randomisation.inf}
\alias{Randomisation.inf}
\title{This function checks the validity of the randomisation information input}
\usage{
Randomisation.inf(Random.inf = Random.inf)
}
\arguments{
\item{Random.inf}{A list of adaptive randomisation information required for randomisation}
}
\value{
#A checked list of input randomisation information
}
\description{
This function checks the validity of the randomisation information input
}
