% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunction.R
\name{simulatetrial}
\alias{simulatetrial}
\title{This function simulate a platform trial applying adaptive methods where the time trend effect can be studied.}
\usage{
simulatetrial(
  ii,
  response.probs = c(0.4, 0.4),
  ns = c(30, 60, 90, 120, 150),
  max.ar = 0.75,
  rand.type = "Urn",
  max.deviation = 3,
  model.inf = list(model = "tlr", ibb.inf = list(pi.star = 0.5, pess = 2,
    betabinomialmodel = ibetabinomial.post), tlr.inf = list(beta0_prior_mu = 0,
    beta1_prior_mu = 0, beta0_prior_sigma = 2.5, beta1_prior_sigma = 2.5, beta0_df = 7,
    beta1_df = 7, reg.inf = "main", variable.inf = "Fixeffect")),
  Stopbound.inf = Stopbound.inf,
  Random.inf = Random.inf,
  trend.inf = trend.inf
)
}
\arguments{
\item{ii}{Meaning less parameter but required for foreach function in doParallel package}

\item{response.probs}{A vector of true response probability for each arm}

\item{ns}{Vector of number of patient at each stage}

\item{max.ar}{The upper boundary for randomisation ratio for each arm}

\item{rand.type}{The randomisation type: "Coin" or "Urn"}

\item{max.deviation}{The tuning parameter for Urn randomisation method}

\item{model.inf}{The list of interim data analysis model information}

\item{Stopbound.inf}{The list of stop boundary information}

\item{Random.inf}{The list of Adaptive randomisation information}

\item{trend.inf}{The list of time trend information}
}
\value{
A matrix including all evaluation metrics
}
\description{
This function simulate a platform trial applying adaptive methods where the time trend effect can be studied.
}
\examples{
set.seed(1)
simulatetrial(response.probs = c(0.4, 0.4),
               ns = c(30, 60, 90, 120, 150),
               max.ar = 0.75,
               rand.type = "Urn",
               max.deviation = 3,
               model.inf = list(
               model = "tlr",
               ibb.inf = list(
               pi.star = 0.5,
               pess = 2,
               betabinomialmodel = ibetabinomial.post
               ),
               tlr.inf = list(
               beta0_prior_mu = 0,
               beta1_prior_mu = 0,
               beta0_prior_sigma = 2.5,
               beta1_prior_sigma = 2.5,
               beta0_df = 7,
               beta1_df = 7,
               reg.inf =  "main",
               variable.inf = "Fixeffect"
               )
               ),
               Stopbound.inf = Stopboundinf(
               Stop.type = "Early-Pocock",
               Boundary.type = "Symmetric",
               cutoff = c(0.99,0.01)
               ),
               Random.inf = list(
               Fixratio = FALSE,
               Fixratiocontrol = NA,
               BARmethod = "Thall",
               Thall.tuning.inf = list(tuningparameter = "Fixed",  fixvalue = 1)
               ),
               trend.inf = list(
               trend.type = "step",
               trend.effect = c(0, 0),
               trend_add_or_multip = "mult"
               ))

#   PP1C nC yC nE1 yE1 H1^1tpIE Intercept Trt1_Mean Trt1_Var
# 1 0.7788 15  4  15   6        0    -1.019     0.569    0.593
# 2 0.8808 22  6  38  16        0    -1.011     0.678    0.336
# 3 0.9040 29  8  61  25        0    -0.990     0.618    0.239
# 4 0.9940 36  8  84  39        1    -1.277     1.124    0.202
# 5     NA NA NA  NA  NA       NA        NA        NA       NA
}
