% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transferdatatolist_Stopboundinf.R
\name{Stopboundinf}
\alias{Stopboundinf}
\title{This function summarises and checks stopping boundary information.}
\usage{
Stopboundinf(
  Stop.type = "Early-Pocock",
  Boundary.type = "Symmetric",
  cutoff = c(0.9928, 0.0072)
)
}
\arguments{
\item{Stop.type}{The shape of stopping boundary}

\item{Boundary.type}{"Symmetric" or "Asymmetric"}

\item{cutoff}{= c(cutoff1, cutoff2) #A vector
#Pr(theta_1 > theta_0|D_n) > cutoff1. Should input the cutoff1 for efficacy boundary as the first element
#Pr(theta_1 < theta_0|D_n) < cutoff2. Should input the cutoff2 for futility boundary as the first element}
}
\value{
The list of information required for boundary construction
}
\description{
This function summarises and checks stopping boundary information.
}
\examples{
Stop.type = "Early-Pocock" #(Pocock boundarty is a flat boundary across time)
Boundary.type = "Symmetric"
cutoff = c(0.9928, 0.0072)
ns = c(60, 120, 180, 240, 300)

Stopbound.inf = Stopboundinf()
#Stopbound.inf
#$Stop.type
# [1] "Early-Pocock"
#$Boundary.type
#[1] "Symmetric"
#$cutoff
# [1] 0.9928 0.0072
#$ns
#[1] 60 120 180 240 300
}
