% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Timetrendfunctiongeneration.R
\name{Timetrend.fun}
\alias{Timetrend.fun}
\title{This function generate the time trend function based on trend information}
\usage{
Timetrend.fun(trend.inf)
}
\arguments{
\item{trend.inf}{#The list of information for time trend effect}
}
\value{
#A list containing the time trend function according to input trend.type variable,
and a indicator of whether there is a time trend in data generation
based on input trend information
}
\description{
This function generate the time trend function based on trend information
}
